requirements.txt
flask==3.0.0
gunicorn>=21.2.0
rembg>=2.0.57
Pillow>=10.2.0
onnxruntime>=1.17.1
scikit-image>=0.21.0
pymatting>=1.1.12
